pyproject.toml
[project]
name = "cs-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphviz>=0.21",
    "langgraph>=1.0.1",
    "openai>=2.6.1",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
]
